Pre-implementation cleanup
