+++ dotnet/private/actions/common.bzl
@@ -1,5 +1,8 @@
 load("//dotnet/private:providers.bzl", "NuGetPackageInfo")
 
+def declare_cache(ctx):
+    return ctx.actions.declare_file(ctx.attr.name + ".cache")
+
 def write_cache_manifest(ctx, caches):
     cache_manifest = ctx.actions.declare_file(ctx.attr.name + ".input_caches")
     ctx.actions.write(cache_manifest, "\n".join([c.path for c in caches.to_list()]))
