+++ dotnet/tools/builder/BazelLogger.cs
@@ -11,7 +11,7 @@ namespace RulesMSBuild.Tools.Builder
             Environment.Exit(1);
             return 1; // weird. Oh well.
         }
-        public static bool DebugEnabled = Environment.GetEnvironmentVariable("DOTNET_BUILDER_DEBUG") != null;
+        public static bool DebugEnabled = Environment.GetEnvironmentVariable("BUILD_DIAG") != null;
         public static bool VerboseEnabled = Environment.GetEnvironmentVariable("DOTNET_BUILDER_DEBUG") == "v";
 
         public static void Verbose(string message)
